+++ Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version:5.0
+// swift-tools-version:5.2
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -30,8 +30,7 @@ let package = Package(
             path: "Sources"),
         .testTarget(
             name: "SBValidatorTests",
-            dependencies: ["SBValidator"],
-            path: "Tests"),
+            dependencies: ["SBValidator"]),
     ],
     swiftLanguageVersions: [.v5]
 )