+++ Tests/LinuxMain.swift
@@ -1,7 +1,8 @@
 import XCTest
 
-import ValidatorTests
+import SBValidator
 
 var tests = [XCTestCaseEntry]()
 tests += ValidatorTests.allTests()
 XCTMain(tests)
+